move b-tree test to its own file
